# genai-engine/staging.openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Arthur GenAI Engine",
        "version": "2.1.0"
    },
    "paths": {
...
        "/api/v2/feedback/{inference_id}": {
            "post": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Post Feedback",
                "description": "Post feedback for LLM Application.",
                "operationId": "post_feedback_api_v2_feedback__inference_id__post",
                "security": [
                    {
                        "API Key": []
                    }
                ],
                "parameters": [
                    {
                        "name": "inference_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Inference Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FeedbackRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InferenceFeedbackResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/chat/files": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Upload Embeddings File",
                "description": "Upload files via form-data. Only PDF, CSV, TXT types accepted.",
                "operationId": "upload_embeddings_file_api_chat_files_post",
                "parameters": [
                    {
                        "name": "is_global",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false,
                            "title": "Is Global"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_embeddings_file_api_chat_files_post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileUploadResult"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Get Files",
                "description": "List uploaded files. Only files that are global or owned by the caller are returned.",
                "operationId": "get_files_api_chat_files_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExternalDocument"
                                    },
                                    "title": "Response Get Files Api Chat Files Get"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/chat/files/{file_id}": {
            "delete": {
                "tags": [
                    "Chat"
                ],
                "summary": "Delete File",
                "description": "Remove a file by ID. This action cannot be undone.",
                "operationId": "delete_file_api_chat_files__file_id__delete",
                "parameters": [
                    {
                        "name": "file_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "File Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/chat/": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Chat",
                "description": "Chat request for Arthur Chat",
                "operationId": "chat_request",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/chat/context/{inference_id}": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Get Inference Document Context",
                "description": "Get document context used for a past inference ID.",
                "operationId": "get_inference_document_context_api_chat_context__inference_id__get",
                "parameters": [
                    {
                        "name": "inference_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Inference Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChatDocumentContext"
                                    },
                                    "title": "Response Get Inference Document Context Api Chat Context  Inference Id  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/chat/feedback/{inference_id}": {
            "post": {
                "tags": [
                    "Chat",
                    "Chat"
                ],
                "summary": "Post Chat Feedback",
                "description": "Post feedback for Arthur Chat.",
                "operationId": "post_chat_feedback_api_chat_feedback__inference_id__post",
                "parameters": [
                    {
                        "name": "inference_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Inference Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FeedbackRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/chat/conversations": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Get Conversations",
                "description": "Get list of conversation IDs.",
                "operationId": "get_conversations_api_chat_conversations_get",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 50,
                            "title": "Size"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Page_List_ConversationBaseResponse__"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/chat/default_task": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Get Default Task",
                "description": "Get the default task for Arthur Chat.",
                "operationId": "get_default_task_api_chat_default_task_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatDefaultTaskResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Chat"
                ],
                "summary": "Update Default Task",
                "description": "Update the default task for Arthur Chat.",
                "operationId": "update_default_task_api_chat_default_task_put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatDefaultTaskRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatDefaultTaskResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Body_upload_embeddings_file_api_chat_files_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_upload_embeddings_file_api_chat_files_post"
            },
...
            "FeedbackRequest": {
                "properties": {
                    "target": {
                        "$ref": "#/components/schemas/InferenceFeedbackTarget"
                    },
                    "score": {
                        "type": "integer",
                        "title": "Score"
                    },
                    "reason": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Reason"
                    },
                    "user_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Id"
                    }
                },
                "type": "object",
                "required": [
                    "target",
                    "score",
                    "reason"
                ],
                "title": "FeedbackRequest"
            },
            "FileUploadResult": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "type": {
                        "type": "string",
                        "title": "Type"
                    },
                    "word_count": {
                        "type": "integer",
                        "title": "Word Count"
                    },
                    "success": {
                        "type": "boolean",
                        "title": "Success"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "type",
                    "word_count",
                    "success"
                ],
                "title": "FileUploadResult"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "InferenceFeedbackResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "inference_id": {
                        "type": "string",
                        "title": "Inference Id"
                    },
                    "target": {
                        "$ref": "#/components/schemas/InferenceFeedbackTarget"
                    },
                    "score": {
                        "type": "integer",
                        "title": "Score"
                    },
                    "reason": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Reason"
                    },
                    "user_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "inference_id",
                    "target",
                    "score",
                    "created_at",
                    "updated_at"
                ],
                "title": "InferenceFeedbackResponse"
            },
...
        },
        "securitySchemes": {
            "API Key": {
                "type": "http",
                "description": "Bearer token authentication with an API key",
                "scheme": "bearer"
            }
        }
    }
}